requirements.txt
openai
elevenlabs==2.8.0
python-dotenv
customtkinter
pillow
PyMuPDF
